feat: create test.util.ts for test code
